+++ include/tao/pegtl/internal/rules.hpp
@@ -9,6 +9,8 @@
 #include "any.hpp"
 #include "apply.hpp"
 #include "apply0.hpp"
+#include "ascii_astring.hpp"
+#include "ascii_aistring.hpp"
 #include "ascii_istring.hpp"
 #include "ascii_string.hpp"
 #include "at.hpp"